WIP: Differentiate tests by integration with unit specificity
And isolate problematic ones

Fixes #989
